backend/requirements.txt
werkzeug
torch
torchvision
timm
opencv-python
numpy
pillow
flask
flask-cors
python-barcode
pandas
